% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gfGapLength}
\alias{gfGapLength}
\title{Julendat function to identify lengths of measurement gaps}
\usage{
gfGapLength(data.dep, pos.na, gap.limit, end.datetime, units, ...)
}
\description{
Julendat function to identify lengths of measurement gaps
}

