\name{gfJulendat}
\alias{gfJulendat}
\title{Fill medium-sized continuous gaps via the Julendat gap-filling routine}
\usage{
gfJulendat(files.dep, files.indep, filepath.output, filepath.coords = NULL,
  quality.levels = NULL, gap.limit, end.datetime, units = "hours", na.limit,
  time.window, n.plot, prm.dep, prm.indep, family, plevel, save.output = F,
  ...)
}
\description{
Fill medium-sized continuous gaps via the Julendat
gap-filling routine
}

