% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gfLinInt}
\alias{gfLinInt}
\title{Fill small continuous gaps via linear interpolation}
\usage{
gfLinInt(data, prm = "TEMP", limit = 5, width = 11, save.output = T,
  ...)
}
\description{
Fill small continuous gaps via linear interpolation
}

