% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gfSsa}
\alias{gfSsa}
\title{Impute missing values based on singular spectrum analysis (SSA)}
\usage{
gfSsa(data, prm = "TEMP", reversed.forecast = F, origin = "1970-01-01",
  save.output = F, ...)
}
\description{
Impute missing values based on singular spectrum analysis (SSA)
}

