\name{gfWrite}
\alias{gfWrite}
\title{Julendat wrapper function for gfRun.R}
\usage{
gfWrite(files.dep, files.indep, filepath.output, filepath.coords = NULL,
  quality.levels = NULL, gap.limit, end.datetime, units = "hours", na.limit,
  time.window, n.plot, prm.dep, prm.indep, family, plevel, ...)
}
\description{
Julendat wrapper function for gfRun.R
}

