\name{gsod2ki}
\alias{gsod2ki}
\title{GSOD to julendat conversion}
\usage{
gsod2ki(file.gsod, comma.separated = T, date.column = "YEARMODA",
  date.start = NULL, date.end = NULL, date.format = "\%Y\%m\%d",
  prm.column = "TEMP", temp2celsius = T, time.step = "day",
  timezone = NA, aggtime = "-999", plot.id = NA, ep.plot.id = "xxx",
  station.id = NA, proc.level = -999, qual.flag = NA, save.output = F,
  ...)
}
\description{
GSOD to julendat conversion
}

