% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsodPlot}
\alias{gsodPlot}
\title{Plot GSOD data}
\usage{
gsodPlot(fls_orig = NULL, fls, stations, prm = "TEMP", type = "trends",
  ...)
}
\arguments{
\item{fls_orig}{Character, default is NULL. Filepath(s) to the initial (or
outlier-adjusted) GSOD data set(s).}

\item{fls}{Character. Filepath(s) to the imputed GSOD data set(s), usually
derived from \code{\link{gfLinInt}}, \code{\link{gfJulendat}} and/or
\code{\link{gfSsa}}.}

\item{stations}{Character. Name(s) of the station(s) corresponding to \code{fls}
and \code{fls_orig} that will be displayed above each facet.}

\item{prm}{Character, default is "TEMP". Determines which parameter to
visualize. If \code{type = "trends"}, this argument will be ignored, and
columns "TEMP", "MIN" and "MAX" must be available in \code{fls} and
\code{fls_orig}.}

\item{...}{Additional arguments passed to \code{\link{element_text}}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
This function aims at visualizing GSOD data, both raw/adjusted for outliers and
imputed, as well as linear trends of mean, minimum and maximum daily air
temperature.
}
\examples{
library(dplyr)

data(gsodstations)
gar <- filter(gsodstations, STATION.NAME == "GARISSA")

gsod_gar <- dlGsodStations(usaf = gar$USAF,
                           start_year = 1990, end_year = 1995,
                           dsn = paste0(getwd(), "/data/gar/"),
                           unzip = TRUE)

# Conversion to KiLi SP1 `ki.data` object
ki_gar <- gsod2ki(data = gsod_gar,
                  prm_col = c("TEMP", "MIN", "MAX"),
                  df2ki = TRUE)

# Fill small gaps (n <= 5) with linear interpolation
ki_gar_linint <- gfLinInt(data = ki_gar,
                          prm = c("TEMP", "MIN", "MAX"))

# Fill remaining gaps based on SSA
ki_gar_ssa <- gfSsa(data = ki_gar_linint,
                    prm = c("TEMP", "MIN", "MAX"),
                    reversed_forecast = FALSE,
                    digits = 2)

plot(slot(ki_gar_ssa, "Parameter")[["TEMP"]], type = "l", col = "red")
lines(slot(ki_gar_linint, "Parameter")[["TEMP"]], col = "green")
lines(slot(ki_gar, "Parameter")[["TEMP"]])
}
\author{
Florian Detsch
}

