% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gsodReformat}
\alias{gsodReformat}
\title{Reformat elevation and coordinates in GSOD stations file}
\usage{
gsodReformat(data, elevation = TRUE, coords = TRUE, df2sp = FALSE, ...)
}
\description{
Reformat elevation and coordinates in GSOD stations file
}

