% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seasonalShift}
\alias{seasonalShift}
\title{Function to plot seasonal shifts in monthly averaged GSOD temperature data}
\usage{
seasonalShift(fls, start = c("1973-01-01", "1977-12-31"),
  end = c("2008-01-01", "2012-12-31"), stations, ncol = 2, cols = NULL,
  ...)
}
\description{
Function to plot seasonal shifts in monthly averaged GSOD temperature data
}

