% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stationFromCoords}
\alias{stationFromCoords}
\title{Identify nearby GSOD stations based on a given geographic location}
\usage{
stationFromCoords(x, y = NULL, width = 50, ...)
}
\description{
Identify nearby GSOD stations based on a given geographic location
}

