% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stationFromExtent}
\alias{stationFromExtent}
\title{Identify stations within user-defined extent from map}
\usage{
stationFromExtent(mapRegion = "world", bb = NULL, ...)
}
\description{
Identify stations within user-defined extent from map
}

