% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gsodPlot}
\alias{gsodPlot}
\title{Function to plot gap-filled GSOD data sets}
\usage{
gsodPlot(fls.orig = NULL, fls.gf, stations, prm.orig = "TEMP",
  type = "filled", ...)
}
\description{
Function to plot gap-filled GSOD data sets
}

